package webservice as docker container: - move files in root dir to flask_app subdir - add Dockerfile / docker-compose yaml - add new setup / build instructions to README
