+++ docker-compose.yaml
@@ -1,0 +1,15 @@
+version: "2"
+
+services:
+  flask_app:
+    image: start_hack
+    build:
+      context: .
+    expose:
+      - 5000
+    network_mode: host
+
+networks:
+  default:
+    external:
+      name: host
