+++ flask_app/static/assets/fonts/boxicons.ttf
