+++ flask_app/static/assets/fonts/boxicons.woff
