+++ flask_app/static/assets/img/example.jpg
