+++ flask_app/static/assets/img/favicon2.ico
