+++ flask_app/static/assets/js/fslightbox.js
