+++ flask_app/static/assets/js/isotope.pkgd.js
