+++ flask_app/static/assets/js/templatemo.js
